added index.phps and ultra basic permissions

+++ dancebuilder/ajax_dancebuilder_delete_move.php
@@ -2,11 +2,13 @@
 namespace dancebuilder;
 define("skip_html", true);
 require_once (__DIR__ . "/../apptop.php");
-$data = \core\Input::GetAll();
-if($data['dance_piece_id'] < 1){
-    echo json_encode(['msg' => "invalid dance_piece_id"]);
-    die();
+if(\core\Permissions::permission_level() == \core\Permissions::admin){
+    $data = \core\Input::GetAll();
+    if($data['dance_piece_id'] < 1){
+        echo json_encode(['msg' => "invalid dance_piece_id"]);
+        die();
+    }
+    
+    \core\DB::run("delete from dance_piece where dance_piece_id = :dance_piece_id", $data);
+    echo json_encode(['msg' => "Dance move successfully deleted"]);
 }
-
-\core\DB::run("delete from dance_piece where dance_piece_id = :dance_piece_id", $data);
-echo json_encode(['msg' => "Dance move successfully deleted"]);
